// src/app/generated_services/model/showTransactionDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransactionType } from './transactionType';


export interface ShowTransactionDTO { 
    id?: string;
    type?: TransactionType;
    transactionCategoryId?: string;
    transactionCategoryName?: string | null;
    description?: string | null;
    amount?: number;
    transactionDate?: string;
    reference?: string | null;
    createdAt?: string;
}
export namespace ShowTransactionDTO {
}